[refactor] Location: detach from() to mapper class with extension function
