+++ src/jvmTest/kotlin/uk/ryanwong/gmap2ics/app/models/timeline/LocationTest.kt
@@ -9,130 +9,6 @@ import io.kotest.matchers.shouldBe
 
 internal class LocationTest : FreeSpec() {
     init {
-        "from Location Data Model" - {
-            "should convert locationDataModel to Location App Model correctly" - {
-                // 🔴 Given
-                val locationDataModel = uk.ryanwong.gmap2ics.data.models.timeline.Location(
-                    address = "some-address",
-                    latitudeE7 = 343970563,
-                    longitudeE7 = 1324677422,
-                    locationConfidence = 0.987654,
-                    name = "some-name",
-                    placeId = "some-place-id"
-                )
-
-                // 🟡 When
-                val locationAppModel = Location.from(locationDataModel = locationDataModel)
-
-                // 🟢 Then
-                locationAppModel shouldBe Location(
-                    placeId = "some-place-id",
-                    latitudeE7 = 343970563,
-                    longitudeE7 = 1324677422,
-                    name = "some-name",
-                    address = "some-address",
-                )
-            }
-
-            "should return null if locationDataModel does not contain latitudeE7" {
-                // 🔴 Given
-                val locationDataModel = uk.ryanwong.gmap2ics.data.models.timeline.Location(
-                    address = "some-address",
-                    latitudeE7 = null,
-                    longitudeE7 = 1324677422,
-                    locationConfidence = 0.987654,
-                    name = "some-name",
-                    placeId = "some-place-id"
-                )
-
-                // 🟡 When
-                val locationAppModel = Location.from(locationDataModel = locationDataModel)
-
-                // 🟢 Then
-                locationAppModel shouldBe null
-            }
-
-            "should return null if locationDataModel does not contain longitudeE7" {
-                // 🔴 Given
-                val locationDataModel = uk.ryanwong.gmap2ics.data.models.timeline.Location(
-                    address = "some-address",
-                    latitudeE7 = 343970563,
-                    longitudeE7 = null,
-                    locationConfidence = 0.987654,
-                    name = "some-name",
-                    placeId = "some-place-id"
-                )
-
-                // 🟡 When
-                val locationAppModel = Location.from(locationDataModel = locationDataModel)
-
-                // 🟢 Then
-                locationAppModel shouldBe null
-            }
-        }
-
-        "from ActivityLocation Data Model" - {
-            "should convert activityLocationDataModel to Location App Model correctly" - {
-                // 🔴 Given
-                val activityLocationDataModel = uk.ryanwong.gmap2ics.data.models.timeline.ActivityLocation(
-                    address = "some-address",
-                    latitudeE7 = 343970563,
-                    longitudeE7 = 1324677422,
-                    locationConfidence = 0.987654,
-                    name = "some-name",
-                    placeId = "some-place-id"
-                )
-
-                // 🟡 When
-                val locationAppModel = Location.from(activityLocationDataModel = activityLocationDataModel)
-
-                // 🟢 Then
-                locationAppModel shouldBe Location(
-                    placeId = "some-place-id",
-                    latitudeE7 = 343970563,
-                    longitudeE7 = 1324677422,
-                    name = "some-name",
-                    address = "some-address",
-                )
-            }
-
-            "should return null if activityLocationDataModel does not contain latitudeE7" {
-                // 🔴 Given
-                val activityLocationDataModel = uk.ryanwong.gmap2ics.data.models.timeline.ActivityLocation(
-                    address = "some-address",
-                    latitudeE7 = null,
-                    longitudeE7 = 1324677422,
-                    locationConfidence = 0.987654,
-                    name = "some-name",
-                    placeId = "some-place-id"
-                )
-
-                // 🟡 When
-                val locationAppModel = Location.from(activityLocationDataModel = activityLocationDataModel)
-
-                // 🟢 Then
-                locationAppModel shouldBe null
-            }
-
-            "should return null if activityLocationDataModel does not contain longitudeE7" {
-                // 🔴 Given
-                val activityLocationDataModel = uk.ryanwong.gmap2ics.data.models.timeline.ActivityLocation(
-                    address = "some-address",
-                    latitudeE7 = 343970563,
-                    longitudeE7 = null,
-                    locationConfidence = 0.987654,
-                    name = "some-name",
-                    placeId = "some-place-id"
-                )
-
-                // 🟡 When
-                val locationAppModel = Location.from(activityLocationDataModel = activityLocationDataModel)
-
-                // 🟢 Then
-                locationAppModel shouldBe null
-            }
-        }
-
         "should return correct formatted values" {
             // 🔴 Given
             val location = Location(
